# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coffee Shop POS API Documentation",
        "description": "Swagger OpenApi description for the Coffee Shop POS system",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Dynamic API Server"
        }
    ],
    "paths": {
...
        "/api/v1/shifts/current": {
            "get": {
                "tags": [
                    "POS Core Data"
                ],
                "summary": "Get Current Shift",
                "description": "Checks if the current cashier has an active shift.",
                "operationId": "ef288be9e78885f1545d59969555786e",
                "responses": {
                    "200": {
                        "description": "Current shift retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "uuid"
                                                },
                                                "opened_at": {
                                                    "type": "string",
                                                    "example": "2023-10-01 08:00:00"
                                                },
                                                "starting_cash": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 150
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/shifts/open": {
            "post": {
                "tags": [
                    "POS Core Data"
                ],
                "summary": "Open Shift",
                "description": "Opens the register for the cashier with a starting float.",
                "operationId": "f0acdd357e345cfba165ed4a3964953d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "opening_cash": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "opening_cash_khr": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 400000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Shift opened successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Shift opened successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "uuid"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/tables": {
            "get": {
                "tags": [
                    "POS Core Data"
                ],
                "summary": "List Tables",
                "description": "Returns all tables and their statuses.",
                "operationId": "845a65ae718789ded7b2648898fb2c8c",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tables retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "uuid"
                                                    },
                                                    "table_number": {
                                                        "type": "string",
                                                        "example": "T1"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "available"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Indoor"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tables/categories": {
            "get": {
                "tags": [
                    "POS Core Data"
                ],
                "summary": "List Table Categories",
                "description": "Returns table zones (e.g., Indoor, Outdoor).",
                "operationId": "336ab22fbc553b9beb6d390af7f35a78",
                "responses": {
                    "200": {
                        "description": "Table categories retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "uuid"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Indoor"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}